Move more code related to finding tests to FindTests.js
