+++ lib/FindTests.js
@@ -1,8 +1,67 @@
 // @flow
 
 const gracefulFs = require('graceful-fs');
+const fs = require('fs');
+const glob = require('glob');
 const path = require('path');
 const Parser = require('./Parser');
+const Project = require('./Project.js');
+
+void Project;
+
+function flatMap/*:: <T, U> */(
+  array /*: Array<T> */,
+  f /*: (T) => Array<U> */
+) /*: Array<U> */ {
+  return array.reduce((result, item) => result.concat(f(item)), []);
+}
+
+// Resolve arguments that look like globs for shells that don’t support globs.
+function resolveGlobs(fileGlobs /*: Array<string> */) /*: Array<string> */ {
+  return flatMap(flatMap(fileGlobs, globify), resolveFilePath);
+}
+
+function globify(globString /*: string */) /*: Array<string> */ {
+  // Without `path.resolve`, `../tests` gives 0 results even if `../tests`
+  // exists (at least on MacOS).
+  return glob.sync(path.resolve(globString), {
+    nocase: true,
+    ignore: '**/elm-stuff/**',
+    nodir: false,
+    absolute: true,
+  });
+}
+
+// Recursively search directories for *.elm files, excluding elm-stuff/
+function resolveFilePath(elmFilePathOrDir /*: string */) /*: Array<string> */ {
+  const candidates = !fs.existsSync(elmFilePathOrDir)
+    ? []
+    : fs.lstatSync(elmFilePathOrDir).isDirectory()
+    ? flatMap(
+        glob.sync('/**/*.elm', {
+          root: elmFilePathOrDir,
+          nocase: true,
+          ignore: '/**/elm-stuff/**',
+          nodir: true,
+        }),
+        resolveFilePath
+      )
+    : [path.resolve(elmFilePathOrDir)];
+
+  // Exclude everything having anything to do with elm-stuff
+  return candidates.filter(
+    (candidate) => !candidate.split(path.sep).includes('elm-stuff')
+  );
+}
+
+function getGlobsToWatch(
+  project /*: typeof Project.Project */
+) /*: Array<string> */ {
+  return project.testsSourceDirs.map((sourceDirectory) =>
+    // TODO: Test this on Windows.
+    path.posix.join(sourceDirectory, '**', '*.elm')
+  );
+}
 
 function findTests(
   testFilePaths /*: Array<string> */,
@@ -118,6 +177,45 @@ Make sure that all parts start with an uppercase letter and don’t contain any 
   `.trim();
 }
 
+function noFilesFoundError(
+  projectRootDir /*: string */,
+  testFileGlobs /*: Array<string> */
+) /*: string */ {
+  return testFileGlobs.length === 0
+    ? `
+${noFilesFoundInTestsDir(projectRootDir)}
+
+To generate some initial tests to get things going: elm-test init
+
+Alternatively, if your project has tests in a different directory,
+try calling elm-test with a glob such as: elm-test "src/**/*Tests.elm"
+      `.trim()
+    : `
+No files found matching:
+
+${testFileGlobs.join('\n')}
+
+Are the above patterns correct? Maybe try running elm-test with no arguments?
+      `.trim();
+}
+
+function noFilesFoundInTestsDir(projectRootDir) {
+  const testsDir = path.join(projectRootDir, 'tests');
+  try {
+    const stats = fs.statSync(testsDir);
+    return stats.isDirectory()
+      ? 'No .elm files found in the tests/ directory.'
+      : `Expected a directory but found something else at: ${testsDir}\nCheck it out! Could you remove it?`;
+  } catch (error) {
+    return error.code === 'ENOENT'
+      ? 'The tests/ directory does not exist.'
+      : `Failed to read the tests/ directory: ${error.message}`;
+  }
+}
+
 module.exports = {
   findTests,
+  getGlobsToWatch,
+  noFilesFoundError,
+  resolveGlobs,
 };
