+++ lib/elm-test.js
@@ -4,7 +4,6 @@ const chalk = require('chalk');
 const chokidar = require('chokidar');
 const { program } = require('commander');
 const fs = require('fs');
-const glob = require('glob');
 const os = require('os');
 const path = require('path');
 const which = require('which');
@@ -19,60 +18,6 @@ const Report = require('./Report.js');
 const Supervisor = require('./Supervisor.js');
 
 void Report;
-
-// Resolve arguments that look like globs for shells that don’t support globs.
-function resolveGlobs(fileGlobs /*: Array<string> */) /*: Array<string> */ {
-  return flatMap(flatMap(fileGlobs, globify), resolveFilePath);
-}
-
-function flatMap/*:: <T, U> */(
-  array /*: Array<T> */,
-  f /*: (T) => Array<U> */
-) /*: Array<U> */ {
-  return array.reduce((result, item) => result.concat(f(item)), []);
-}
-
-function globify(globString /*: string */) /*: Array<string> */ {
-  // Without `path.resolve`, `../tests` gives 0 results even if `../tests`
-  // exists (at least on MacOS).
-  return glob.sync(path.resolve(globString), {
-    nocase: true,
-    ignore: '**/elm-stuff/**',
-    nodir: false,
-    absolute: true,
-  });
-}
-
-// Recursively search directories for *.elm files, excluding elm-stuff/
-function resolveFilePath(elmFilePathOrDir /*: string */) /*: Array<string> */ {
-  const candidates = !fs.existsSync(elmFilePathOrDir)
-    ? []
-    : fs.lstatSync(elmFilePathOrDir).isDirectory()
-    ? flatMap(
-        glob.sync('/**/*.elm', {
-          root: elmFilePathOrDir,
-          nocase: true,
-          ignore: '/**/elm-stuff/**',
-          nodir: true,
-        }),
-        resolveFilePath
-      )
-    : [path.resolve(elmFilePathOrDir)];
-
-  // Exclude everything having anything to do with elm-stuff
-  return candidates.filter(
-    (candidate) => !candidate.split(path.sep).includes('elm-stuff')
-  );
-}
-
-function getGlobsToWatch(
-  project /*: typeof Project.Project */
-) /*: Array<string> */ {
-  return project.testsSourceDirs.map((sourceDirectory) =>
-    // TODO: Test this on Windows.
-    path.posix.join(sourceDirectory, '**', '*.elm')
-  );
-}
 
 // Incorporate the process PID into the socket name, so elm-test processes can
 // be run parallel without accidentally sharing each others' sockets.
@@ -204,17 +149,19 @@ function test(
         queue.length > 0 && queue.every(({ event }) => event === 'changed');
       queue = [];
       if (!onlyChanged) {
-        testFilePaths = resolveGlobs(
+        testFilePaths = FindTests.resolveGlobs(
           testFileGlobs.length === 0 ? [project.testsDir] : testFileGlobs
         );
       }
 
       if (testFilePaths.length === 0) {
-        throw new Error(noFilesFoundError(project.rootDir, testFileGlobs));
+        throw new Error(
+          FindTests.noFilesFoundError(project.rootDir, testFileGlobs)
+        );
       }
 
       if (watcher !== undefined) {
-        const nextGlobsToWatch = getGlobsToWatch(project);
+        const nextGlobsToWatch = FindTests.getGlobsToWatch(project);
         const diff = diffArrays(watchedGlobs, nextGlobsToWatch);
         watchedGlobs = nextGlobsToWatch;
         watcher.add(diff.added);
@@ -324,39 +271,6 @@ function test(
     currentRun = run().then(onRunFinish);
   } else {
     run();
-  }
-}
-
-function noFilesFoundError(projectRootDir, testFileGlobs) {
-  return testFileGlobs.length === 0
-    ? `
-${noFilesFoundInTestsDir(projectRootDir)}
-
-To generate some initial tests to get things going: elm-test init
-
-Alternatively, if your project has tests in a different directory,
-try calling elm-test with a glob such as: elm-test "src/**/*Tests.elm"
-      `.trim()
-    : `
-No files found matching:
-
-${testFileGlobs.join('\n')}
-
-Are the above patterns correct? Maybe try running elm-test with no arguments?
-      `.trim();
-}
-
-function noFilesFoundInTestsDir(projectRootDir) {
-  const testsDir = path.join(projectRootDir, 'tests');
-  try {
-    const stats = fs.statSync(testsDir);
-    return stats.isDirectory()
-      ? 'No .elm files found in the tests/ directory.'
-      : `Expected a directory but found something else at: ${testsDir}\nCheck it out! Could you remove it?`;
-  } catch (error) {
-    return error.code === 'ENOENT'
-      ? 'The tests/ directory does not exist.'
-      : `Failed to read the tests/ directory: ${error.message}`;
   }
 }
 
@@ -578,7 +492,7 @@ function main() {
       const project = getProject('make');
       Generate.generateElmJson(project);
       Compile.compileSources(
-        resolveGlobs(testFileGlobs),
+        FindTests.resolveGlobs(testFileGlobs),
         project.generatedCodeDir,
         pathToElmBinary,
         options.report
